Share Colab dependency detection
