+++ .github/workflows/vignette-artifacts.yml
@@ -12,6 +12,7 @@ on:
       - tools/vignette_artifacts.R
       - tools/render-vignette-pdf.R
       - tools/build-colab-notebook.py
+      - tools/colab_dependencies.py
       - tools/test-vignette-artifacts.py
       - vignettes/**
